@@ -1,7 +1,7 @@
 package com.yahoo.imapnio.async.request;
 
 import javax.annotation.Nonnull;
-import javax.mail.Flags;
+import jakarta.mail.Flags;
 
 import com.yahoo.imapnio.async.data.MessageNumberSet;
 
