@@ -4,8 +4,8 @@ import java.nio.charset.StandardCharsets;
 
 import javax.annotation.Nonnull;
 
-import com.sun.mail.imap.protocol.BASE64MailboxEncoder;
-import com.sun.mail.imap.protocol.MessageSet;
+import org.eclipse.angus.mail.imap.protocol.BASE64MailboxEncoder;
+import org.eclipse.angus.mail.imap.protocol.MessageSet;
 import com.yahoo.imapnio.async.data.MessageNumberSet;
 import com.yahoo.imapnio.async.exception.ImapAsyncClientException;
 
