@@ -4,7 +4,7 @@ import java.nio.charset.StandardCharsets;
 
 import javax.annotation.Nonnull;
 
-import com.sun.mail.imap.protocol.UIDSet;
+import org.eclipse.angus.mail.imap.protocol.UIDSet;
 import com.yahoo.imapnio.async.data.MessageNumberSet;
 
 import io.netty.buffer.ByteBuf;
