@@ -5,9 +5,9 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.util.Properties;
 
-import com.sun.mail.iap.Protocol;
-import com.sun.mail.iap.ProtocolException;
-import com.sun.mail.util.MailLogger;
+import org.eclipse.angus.mail.iap.Protocol;
+import org.eclipse.angus.mail.iap.ProtocolException;
+import org.eclipse.angus.mail.util.MailLogger;
 
 /**
  * Class used to serialize IMAP commands. kraman
